+++ pkg/k8s/scaledobject.go
@@ -1,6 +1,8 @@
 package k8s
 
 import (
+	"strings"
+
 	kedav1alpha1 "github.com/kedacore/keda/v2/apis/keda/v1alpha1"
 	appsv1 "k8s.io/api/apps/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -12,7 +14,7 @@ const (
 	soTriggerType     = "external-push"
 
 	mkScalerAddress = "scalerAddress"
-	mkHost          = "host"
+	mkHosts         = "hosts"
 )
 
 // NewScaledObject creates a new ScaledObject in memory
@@ -21,7 +23,7 @@ func NewScaledObject(
 	name string,
 	deploymentName string,
 	scalerAddress string,
-	host string,
+	hosts []string,
 	minReplicas *int32,
 	maxReplicas *int32,
 	cooldownPeriod *int32,
@@ -54,7 +56,7 @@ func NewScaledObject(
 					Type: soTriggerType,
 					Metadata: map[string]string{
 						mkScalerAddress: scalerAddress,
-						mkHost:          host,
+						mkHosts:         strings.Join(hosts, ","),
 					},
 				},
 			},