+++ scaler/handlers.go
@@ -22,7 +22,8 @@ func init() {
 }
 
 const (
-	interceptor = "interceptor"
+	interceptor  = "interceptor"
+	httpRequests = "http-requests"
 )
 
 type impl struct {
@@ -59,32 +60,38 @@ func (e *impl) IsActive(
 	scaledObject *externalscaler.ScaledObjectRef,
 ) (*externalscaler.IsActiveResponse, error) {
 	lggr := e.lggr.WithName("IsActive")
-	host, ok := scaledObject.ScalerMetadata["host"]
-	if !ok {
-		err := fmt.Errorf("no 'host' field found in ScaledObject metadata")
-		lggr.Error(err, "returning immediately from IsActive RPC call", "ScaledObject", scaledObject)
+
+	hosts, err := getHostsFromScaledObjectRef(lggr, scaledObject)
+	if err != nil {
 		return nil, err
-	}
-	if host == interceptor {
-		return &externalscaler.IsActiveResponse{
-			Result: true,
-		}, nil
 	}
 
-	hostCount, ok := getHostCount(
-		host,
-		e.pinger.counts(),
-		e.routingTable,
-	)
-	if !ok {
-		err := fmt.Errorf("host '%s' not found in counts", host)
-		allCounts := e.pinger.mergeCountsWithRoutingTable(
+	totalHostCount := 0
+	for _, host := range hosts {
+		if host == interceptor {
+			return &externalscaler.IsActiveResponse{
+				Result: true,
+			}, nil
+		}
+
+		hostCount, ok := getHostCount(
+			host,
+			e.pinger.counts(),
 			e.routingTable,
 		)
-		lggr.Error(err, "Given host was not found in queue count map", "host", host, "allCounts", allCounts)
-		return nil, err
+		if !ok {
+			err := fmt.Errorf("host '%s' not found in counts", host)
+			allCounts := e.pinger.mergeCountsWithRoutingTable(
+				e.routingTable,
+			)
+			lggr.Error(err, "Given host was not found in queue count map", "host", host, "allCounts", allCounts)
+			return nil, err
+		}
+
+		totalHostCount += hostCount
 	}
-	active := hostCount > 0
+
+	active := totalHostCount > 0
 	return &externalscaler.IsActiveResponse{
 		Result: active,
 	}, nil
@@ -131,13 +138,15 @@ func (e *impl) GetMetricSpec(
 	sor *externalscaler.ScaledObjectRef,
 ) (*externalscaler.GetMetricSpecResponse, error) {
 	lggr := e.lggr.WithName("GetMetricSpec")
-	host, ok := sor.ScalerMetadata["host"]
-	if !ok {
-		err := fmt.Errorf("'host' not found in ScaledObject metadata")
-		lggr.Error(err, "no 'host' found in ScaledObject metadata")
+
+	hosts, err := getHostsFromScaledObjectRef(lggr, sor)
+	if err != nil {
 		return nil, err
 	}
+
 	var targetPendingRequests int64
+	var host = hosts[0] // We are only interested in the first host to get the targetPendingRequests
+
 	if host == interceptor {
 		targetPendingRequests = e.targetMetricInterceptor
 	} else {
@@ -151,15 +160,16 @@ func (e *impl) GetMetricSpec(
 			)
 			return nil, err
 		}
+		host = httpRequests
 		targetPendingRequests = int64(target.TargetPendingRequests)
 	}
+
 	metricSpecs := []*externalscaler.MetricSpec{
 		{
 			MetricName: host,
 			TargetSize: targetPendingRequests,
 		},
 	}
-
 	return &externalscaler.GetMetricSpecResponse{
 		MetricSpecs: metricSpecs,
 	}, nil
@@ -170,34 +180,40 @@ func (e *impl) GetMetrics(
 	metricRequest *externalscaler.GetMetricsRequest,
 ) (*externalscaler.GetMetricsResponse, error) {
 	lggr := e.lggr.WithName("GetMetrics")
-	host, ok := metricRequest.ScaledObjectRef.ScalerMetadata["host"]
-	if !ok {
-		err := fmt.Errorf("no 'host' field found in ScaledObject metadata")
-		lggr.Error(err, "ScaledObjectRef", metricRequest.ScaledObjectRef)
+
+	hosts, err := getHostsFromScaledObjectRef(lggr, metricRequest.ScaledObjectRef)
+	if err != nil {
 		return nil, err
 	}
 
-	hostCount, ok := getHostCount(
-		host,
-		e.pinger.counts(),
-		e.routingTable,
-	)
-	if !ok {
-		if host == interceptor {
+	var totalCount int64
+	var metricName = httpRequests
+	for _, host := range hosts {
+		hostCount, ok := getHostCount(
+			host,
+			e.pinger.counts(),
+			e.routingTable,
+		)
+		if !ok {
+			if host != interceptor {
+				err := fmt.Errorf("host '%s' not found in counts", host)
+				allCounts := e.pinger.mergeCountsWithRoutingTable(e.routingTable)
+				lggr.Error(err, "allCounts", allCounts)
+				return nil, err
+			}
 			hostCount = e.pinger.aggregate()
-		} else {
-			err := fmt.Errorf("host '%s' not found in counts", host)
-			allCounts := e.pinger.mergeCountsWithRoutingTable(e.routingTable)
-			lggr.Error(err, "allCounts", allCounts)
-			return nil, err
+			metricName = interceptor
 		}
+		totalCount += int64(hostCount)
 	}
+
 	metricValues := []*externalscaler.MetricValue{
 		{
-			MetricName:  host,
-			MetricValue: int64(hostCount),
+			MetricName:  metricName,
+			MetricValue: totalCount,
 		},
 	}
+
 	return &externalscaler.GetMetricsResponse{
 		MetricValues: metricValues,
 	}, nil